Refactor: Create service for technology business logic, refactor technology controller
