+++ server/controllers/technology.controller.js
@@ -1,8 +1,8 @@
-const Technology = require('../models/Technology')
+const { getTechnologies } = require('../services/technology.service')
 
 exports.getTechnologies = async (req, res, next) => {
   try {
-    const technologies = await Technology.find().sort('name')
+    const technologies = await getTechnologies()
     res.status(200).send(technologies)
   } catch (error) {
     next(error)
