+++ server/index.js
@@ -27,7 +27,7 @@ const offers = require('./routes/offers.route')
 const auth = require('./routes/auth.route')
 const products = require('./routes/products.route')
 const users = require('./routes/users')
-const technologies = require('./routes/technologies')
+const technologies = require('./routes/technologies.route')
 
 app.use('/api/v1/offers', offers)
 app.use('/api/v1/auth', auth)
