+++ server/routes/technologies.route.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 
-const { getTechnologies } = require('../controllers/technologies')
+const { getTechnologies } = require('../controllers/technology.controller')
 
 router.get('/', getTechnologies)
